[BUGFIX] Check hotspots when unfolding
This fixes the case when a document is not lexed,
user folds all elements, then expands one lower down.

If a hyperlink is embedded in this unfolded block, it will
not have been styled. Therefore, we check when expanding the fold, but ignore the commands when
collapsing all elements (for performance).