Replace function now working
Issue was with replace needing to "double-check" that the
selection still matches, and therefore needs to force to
allow an empty match at start point. It happened to work
previously when clickable links was enabled, as between
each replace was a search to check if any clickable links
had been introduced, and therefore the "replace" search
was never flagged as a continuation search by the
boostregexsearch class.